Move non-provider specific upgrade tests logic to upgrades package
